Updated error handling and tests

+++ controllers/comment-controller.js
@@ -1,35 +1,39 @@
-const mongoose = require('mongoose')
-const { Article, ArticleComment, Topic, User } = require('../models');
+const mongoose = require("mongoose");
+const { Article, ArticleComment, Topic, User } = require("../models");
 
 exports.updateCommentVotes = (req, res, next) => {
-  const {comment_id} = req.params;
-  const {vote} = req.query
-  let voteChange = 0;
-  if(vote === 'up') voteChange++
-  if(vote === 'down') voteChange--
+  const { comment_id } = req.params;
 
-  return ArticleComment.findByIdAndUpdate(comment_id, {$inc: {votes: voteChange}}, {new: true})
-  .then(comment => {
-    res.status(201).send(comment)
-  })
-  .catch(err => {
-    if(err.name === 'CastError') next({status: 404})
-    else next({status: 500})
-  })
-}
+  if (!mongoose.Types.ObjectId.isValid(comment_id))
+    return next({ status: 400 });
+
+  const { vote } = req.query;
+  let voteChange = 0;
+  if (vote === "up") voteChange++;
+  if (vote === "down") voteChange--;
+
+  return ArticleComment.findByIdAndUpdate(
+    comment_id,
+    { $inc: { votes: voteChange } },
+    { new: true }
+  )
+    .then(comment => {
+      if (!comment) return next({ status: 404 });
+      res.status(201).send(comment);
+    })
+    .catch(next);
+};
 
 exports.deleteComment = (req, res, next) => {
-  const {comment_id} = req.params;
+  const { comment_id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(comment_id))
+    return next({ status: 400 });
+
   ArticleComment.findByIdAndRemove(comment_id)
-  .then( removedComment => {
-    if(!removedComment) return next({status:404})
-    res.status(204).send({})
-  })
-  .catch(err => {
-    if(err.name === 'CastError') next({status: 404})
-    else next({status: 500})
-  })
-}
-
-
-
+    .then(removedComment => {
+      if (!removedComment) return next({ status: 404 });
+      res.status(204).send({});
+    })
+    .catch(next);
+};
